moved the Object calss from Rendering namespace to Core
